sep TDB to TDB and TDBA, update parseBytes fun, add tdb_split opc
